+++ admin/manage_projects.php
@@ -316,8 +316,11 @@ function statusClass($s){
                                 <?php endif; ?>
                                 
                                 <!-- Delete Button -->
-                                <a href="delete_project.php?id=<?= (int)$r['id'] ?>" class="btn btn-sm btn-outline-danger" title="Delete"
-                                   onclick="return confirm('Are you sure you want to delete this activity? This action cannot be undone.');">
+                                <a href="delete_project.php?id=<?= (int)$r['id'] ?>" class="btn btn-sm btn-outline-danger js-delete-confirm" title="Delete"
+                                   data-delete-title="Delete Activity"
+                                   data-delete-message="Are you sure you want to delete this activity?"
+                                   data-delete-item="<?= h($r['project_name'] ?? ('Activity #' . (int)$r['id'])) ?>"
+                                   data-delete-confirm-label="<i class='fa-solid fa-trash me-1'></i> Delete Activity">
                                     <i class="fa-solid fa-trash"></i>
                                     <span class="d-none d-md-inline ms-1">Delete</span>
                                 </a>
